REFERENCE-FOR-REVIEWERS-VMWG_NIST_SP1500-10X-DRAFT_20170824_LML.pdf
REFERENCE-FOR-REVIEWERS-VMWG_NIST_SP1500-10X-DRAFT_20170824_LML.pdf 
This is the draft of the NIST_SP1500-10X standard spec. that the Oct. 2017 
"3.1.4.4 Multi-round Voting Tabulation Methods :  Ranked-Choice-Voting" will become a section of.
Note that SP1500-10X Voting Methods Models is NOT under review at this time, it is intended to provide context and will be offered for review and comment at an upcoming date in Q4 2017. For now, please do not enter comments on SP1500-10X Voting Methods Models in the "3.1.4.4 Multi-round Voting Tabulation Methods :  Ranked-Choice-Voting" Reviewer Comment Spreadsheet! TY. -LL Editor