REVIEW-NIST-1500-10X-VMWG_RCVFlow_10-18-2017_HD-DC-GG-CH_LL-EditorDraft.pdf
This is the copy for review of
3.1.4.4 Multi-round Voting Tabulation Methods :  Ranked-Choice-Voting Oct. 2017
Please only use this .pdf copy for reviewing.  Other files/copies of the document do not have the requisite line numbering you will need to enter comments or change requests here:
 https://docs.google.com/spreadsheets/d/1u2K-OadbGfdEXhfo0tXL65DSORhvC2f_bqo2QLEPt4M/edit#gid=0
(and other copies of the document may be prior versions and not the current version). 
REVIEW-NIST-1500-10X-VMWG_RCVFlow_10-18-2017_HD-DC-GG-CH_LL-EditorDraft.pdf