REVIEW-NIST-1500-10X-VMWG_RCVFlow_Section3144-10-15-2017_HD-DC-GG-CH_LL-EditorDraft.pdf
This is the document for review 10/17. It is a DRAFT section of draft NIST-SP1500-10X Voting Methods.
This draft section is one of the process flow models, with a detailed technical description of how the process flow model provides coverage for the most common voting method variants in its category  (and with intended upcoming UML class diagrams illustrating example CDF's XML/JSON use). It is one of the NIST-SP1500-10X Voting Methods Section 3 artifacts. The Section 3 artifacts are the result of applying a knowledge acquisition and representation methodology to produce an intermediate step in modeling the domain of voting methods in-use and more specifically as in-use and as implemented in practice and according to statute and local rules, in the United States.  Each voting method variant that has a domain model and process flow entry in Section 3., corresponds to an (upcoming) entry in Section 4. where the actual elements of the standard, mechanized (executable) mathematical models written in a human-readable Domain Specific Language (DSL).
Several other DRAFT artifacts are relevant to this review, as context or background information, and are listed in the Invitation to Review as well as listed in the https://github.com/usnistgov/VotingMethods 'readme' / front page of this GitHub repository.
REVIEW-NIST-1500-10X-VMWG_RCVFlow_Section3144-10-15-2017_HD-DC-GG-CH_LL-EditorDraft.pdf -LL